Update README.md
This repository, called "LeetCode," provides as a platform for recording everyday solutions to LeetCode difficulties. Every day, a new challenge gets solved and committed to the repository. This repository, which maintains a constant practice of problem solving, not only chronicles personal improvement but also provides as a great resource for those interested in algorithmic problem solving. Navigating through solutions is simple because to clear commit messages and orderly directory structures. Whether for personal development or collaborative learning, "LeetCode" provides a collection of varied problem-solving methodologies and insights that contribute to the knowledge exchange of the larger programming community.